ref(sight_search_screen): refactoring screen/sight_search_screen
